refactor funding project areas selector
