+++ src/interface/src/app/funding/full-report-view/full-report-view.component.ts
@@ -16,7 +16,11 @@ import { NavBarComponent } from '@app/standalone/nav-bar/nav-bar.component';
 import { ProjectArea } from '@app/types';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
 import { FundingReportService } from '@services/funding-report.service';
-import { FilterDropdownComponent, OpacitySliderComponent } from '@styleguide';
+import { OpacitySliderComponent } from '@styleguide';
+import {
+  FilterProjectFormat,
+  FundingProjectAreasSelectorComponent,
+} from '../funding-project-areas-selector/funding-project-areas-selector.component';
 import {
   ToggleButtonsConfig,
   ToggleTabsComponent,
@@ -46,12 +50,6 @@ import { ScenarioService } from '@app/services';
 import { isPlanningApproachSubUnits } from '@app/scenario/scenario-helper';
 import { FUNDING_REPORT_INFO_URL } from '@shared';
 
-export interface FilterProjectFormat {
-  id: number;
-  name: string;
-  shortName: string;
-}
-
 @UntilDestroy()
 @Component({
   selector: 'app-full-report-view',
@@ -60,7 +58,7 @@ export interface FilterProjectFormat {
     AsyncPipe,
     BaseLayersComponent,
     DataLayersComponent,
-    FilterDropdownComponent,
+    FundingProjectAreasSelectorComponent,
     FundingReportComponent,
     NgIf,
     MatButtonToggleModule,
@@ -69,8 +67,6 @@ export interface FilterProjectFormat {
     MatProgressSpinnerModule,
     MatSelectModule,
     MatTabsModule,
-    FilterDropdownComponent,
-    FundingReportComponent,
     FundingReportMapComponent,
     MapNavbarComponent,
     NavBarComponent,
@@ -131,14 +127,6 @@ export class FullReportViewComponent implements OnInit {
     }),
     shareReplay(1)
   );
-
-  readonly filteredProjectAreas$: Observable<FilterProjectFormat[]> =
-    combineLatest(this.filterOptions$, this.selectedProjectAreas$).pipe(
-      map(([available, selectedIds]) => {
-        if (!selectedIds?.length) return [];
-        return available.filter((area) => selectedIds.includes(area.id));
-      })
-    );
 
   /**
    * Single (non-polling) fetch of the report. This view is only reachable for a
@@ -210,11 +198,6 @@ export class FullReportViewComponent implements OnInit {
         untilDestroyed(this)
       )
       .subscribe(() => this.router.navigate(['/home']));
-  }
-
-  handleFilterSelection(selectedAreas: FilterProjectFormat[]): void {
-    const ids = selectedAreas.map((a) => a.id);
-    this.fundingMapConfigState.updateSelectedProjectAreas(ids);
   }
 
   handleToggleSelection(selection: string): void {
