+++ src/interface/src/app/funding/funding-project-areas-selector/funding-project-areas-selector.component.html
@@ -1,0 +1,19 @@
+<div class="view-outcomes-row">
+  <div class="view-outcomes-box">
+    <div>Viewing outcomes for:</div>
+    <sg-filter-dropdown
+      class="outcome-areas-dropdown"
+      size="medium"
+      (confirmedSelection)="handleFilterSelection($event)"
+      [styleType]="'white'"
+      [hasSearch]="false"
+      displayField="name"
+      shortLabel="shortName"
+      noSelectionsLabel="All Project Areas"
+      [selectedItems]="(filteredProjectAreas$ | async) ?? []"
+      [menuItems]="(filterOptions$ | async) ?? []"
+      [showCountChip]="false"
+      menuLabel="Project Areas">
+    </sg-filter-dropdown>
+  </div>
+</div>
